// src/app/api/models/shoe-care.ts
/* tslint:disable */
/* eslint-disable */
export interface ShoeCare {

  /**
   * The brand of the shoe care
   */
  brand: string;

  /**
   * The unique ID of the shoe care
   */
  id?: number;

  /**
   * The name of the shoe care
   */
  name: string;

  /**
   * The price of the shoe care
   */
  price: number;
}